+++ app/build.gradle
@@ -5,7 +5,7 @@ android {
     buildToolsVersion "30.0.0"
 
     defaultConfig {
-        applicationId "com.example.android.wordgame"
+        applicationId "com.alnamaa.arabic_quiz"
         minSdkVersion 19
         targetSdkVersion 29
         versionCode 1
