+++ app/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.example.android.wordgame">
+    package="com.alnamaa.arabic_quiz">
 
     <application
         android:allowBackup="true"
@@ -10,15 +10,15 @@
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
 
-        <activity android:name=".ui.MainActivity">
+        <activity android:name="com.alnamaa.arabic_quiz.ui.MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".ui.levelsListScreen.LevelsActivity"></activity>
-        <activity android:name=".ui.QuizScreen.QuizActivity"></activity>
+        <activity android:name="com.alnamaa.arabic_quiz.ui.levelsListScreen.LevelsActivity"></activity>
+        <activity android:name="com.alnamaa.arabic_quiz.ui.QuizScreen.QuizActivity"></activity>
     </application>
 
 </manifest>
