--- app/src/main/java/com/alnamaa/arabic_quiz/MainActivity.java
@@ -1,17 +1,0 @@
-package com.alnamaa.arabic_quiz;
-
-import androidx.appcompat.app.AppCompatActivity;
-
-import android.os.Bundle;
-import android.widget.Toast;
-
-public class MainActivity extends AppCompatActivity {
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_main);
-
-        Toast.makeText(this, "Hello World2!", Toast.LENGTH_SHORT).show();
-    }
-}
