+++ app/src/main/java/com/alnamaa/arabic_quiz/models/Question.java
@@ -1,4 +1,4 @@
-package com.example.android.wordgame.models;
+package com.alnamaa.arabic_quiz.models;
 
 import java.util.Arrays;
 import java.util.Collections;
