+++ app/src/main/java/com/alnamaa/arabic_quiz/models/QuizLevel.java
@@ -1,4 +1,4 @@
-package com.example.android.wordgame.models;
+package com.alnamaa.arabic_quiz.models;
 
 public class QuizLevel {
     private int totalScore;
