+++ app/src/main/java/com/alnamaa/arabic_quiz/ui/MainActivity.java
@@ -1,4 +1,4 @@
-package com.example.android.wordgame.ui;
+package com.alnamaa.arabic_quiz.ui;
 
 import androidx.appcompat.app.AppCompatActivity;
 
@@ -7,8 +7,8 @@ import android.os.Bundle;
 import android.view.View;
 import android.widget.Button;
 
-import com.example.android.wordgame.R;
-import com.example.android.wordgame.ui.levelsListScreen.LevelsActivity;
+import com.alnamaa.arabic_quiz.R;
+import com.alnamaa.arabic_quiz.ui.levelsListScreen.LevelsActivity;
 
 public class MainActivity extends AppCompatActivity {
